Move body generator for externally tagged enums to its own module
Cut-paste
